Moved stack using queue to others
